'orders_protected'

+++ back/routes/users.js
@@ -103,7 +103,7 @@ router.post('/login', async (req, res) => {
                     }, iat: Math.floor(Date.now() / 1000) + (60 * 60),
                 }, 'token', (err, token) => {
                     res.status(200).json({
-                       user: user, token : token
+                        user: user, token: token
                     });
                 });
             } else {
@@ -129,16 +129,61 @@ router.put("/:id", async (req, res) => {
     }
 });
 
-router.delete("/:id", async (req, res) => {
-    try {
-        await client.connect();
-        const query = { _id: new ObjectId(req.params.id) };
-        await userCol.deleteOne(query);
-        res.send("Successfully deleted!");
-    } finally {
-        await client.close();
-    }
+router.delete("/:id", verifyToken, async (req, res) => {
+    jwt.verify(req.token, 'token', async (err, authData) => {
+        if (authData.user.role === 'Serveur') {
+            res.status(403).send('Accès interdit')
+        } else {
+            try {
+                await client.connect();
+                const query = { _id: new ObjectId(req.params.id) };
+                await userCol.deleteOne(query);
+                res.send("Successfully deleted!");
+            } finally {
+                await client.close();
+            }
+        }
+    });
+    //Vue profil
+
+
+    router.post('/profile', verifyToken, async (req, res) => {
+        jwt.verify(req.token, 'token', async (err, authData) => {
+            try {
+                await client.connect();
+                const query = { _id: new ObjectId(authData.user.id) };
+                const users = await userCol.findOne(query)
+                res.send(authData);
+            } finally {
+                await client.close();
+            }
+        })
+
+    })
 });
 
-module.exports = router;
+//On verifie tokenn
+function verifyToken(req, res, next) {
+    // Get auth header value
+    const bearerHeader = req.headers['authorization'];
+    // Check if bearer is undefined
+    if (typeof bearerHeader !== 'undefined') {
+        // Split at the space
+        const bearer = bearerHeader.split(' ');
+        // Get token from array
+        const bearerToken = bearer[1];
+        // Set the token
+        req.token = bearerToken;
+        // Next middleware
+        next();
+    } else {
+        // Forbidden
+        res.sendStatus(403);
+    }
+}
+
+module.exports = {
+    router: router,
+    verifyToken: verifyToken
+};
